refactor(eslint): group local rules by rule
