+++ eslint-local-rules/analytics-event-name/rule.ts
